Fixeado

+++ backend/src/main/java/com/alquileres/repository/EstadoInmuebleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EstadoInmuebleRepository extends JpaRepository<EstadoInmueble, Long> {
+public interface EstadoInmuebleRepository extends JpaRepository<EstadoInmueble, Integer> {
     Optional<EstadoInmueble> findByNombre(String nombre);
 }
